Keyboard accessibility
To make the page keyboard accessible without using any JavaScript, we
need to be smart about which elements are focusable at any given time.

Initially only #open is focusable. When this element is focused, the
envelope is outlined.

When #open is checked, #create becomes focusable. When #create is
focused and unchecked, the "Create own letter" button is outlined. When
#create is focused and checked, the "Cancel" button is outlined.

To prevent users being able to tab to the input fields inside of the
dialog, when the dialog is closed, these elements are initially marked
as unfocusable. Only when #create is checked do they become focusable.

To make elements unfocusable, `display: none` is used. This has the
obvious drawback of causing elements to not be displayed anymore. To
work around this, "fake" static elements are unhidden when the primary
interactive elements get hidden.

Finally, we want to prevent users from closing the envelope without
first closing the dialog. Without using JavaScript, this is only
possible in browsers that support :has (Safari). This is done by making
#open unfocusable when #create is checked. For browsers that do not
support has, users can use keyboard navigation to close the envelope
without first closing the dialog. To provide a graceful fallback for
these browsers, the dialog has been updated to automatically close with
the envelope when it is closed.